requirements.txt
mcp
pypdf
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
llama-parse
python-dotenv
nest-asyncio
neo4j>=5.0.0
google-genai